Move JadConfig StringMapConverter into common core package
