+++ metrics-reporter-core/README.md
@@ -1,0 +1,3 @@
+# metrics-reporter-core
+
+Shared core classes for Graylog metrics reporter plugins.
