+++ metrics-reporter-core/src/test/java/org/graylog/plugins/metrics/core/jadconfig/StringMapConverterTest.java
@@ -1,24 +1,23 @@
 /**
- * This file is part of Graylog Metrics Elasticsearch Reporter Plugin.
+ * This file is part of Graylog Metrics Reporter Core Classes.
  *
- * Graylog Metrics Elasticsearch Reporter Plugin is free software: you can redistribute it and/or modify
+ * Graylog Metrics Reporter Core Classes is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
  * the Free Software Foundation, either version 3 of the License, or
  * (at your option) any later version.
  *
- * Graylog Metrics Elasticsearch Reporter Plugin is distributed in the hope that it will be useful,
+ * Graylog Metrics Reporter Core Classes is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU General Public License for more details.
  *
  * You should have received a copy of the GNU General Public License
- * along with Graylog Metrics Elasticsearch Reporter Plugin.  If not, see <http://www.gnu.org/licenses/>.
+ * along with Graylog Metrics Reporter Core Classes.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.graylog.plugins.metrics.elasticsearch.jadconfig;
+package org.graylog.plugins.metrics.core.jadconfig;
 
 import com.github.joschi.jadconfig.ParameterException;
 import com.google.common.collect.ImmutableMap;
-import org.graylog.plugins.metrics.elasticsearch.jadconfig.StringMapConverter;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
