+++ metrics-reporter-elasticsearch/pom.xml
@@ -19,6 +19,11 @@
 
     <dependencies>
         <dependency>
+            <groupId>org.graylog.plugins</groupId>
+            <artifactId>metrics-reporter-core</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+        <dependency>
             <groupId>org.elasticsearch</groupId>
             <artifactId>metrics-elasticsearch-reporter</artifactId>
             <version>2.2.0</version>
