+++ metrics-reporter-influxdb/pom.xml
@@ -23,6 +23,11 @@
 
     <dependencies>
         <dependency>
+            <groupId>org.graylog.plugins</groupId>
+            <artifactId>metrics-reporter-core</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+        <dependency>
             <groupId>com.izettle</groupId>
             <artifactId>metrics-influxdb</artifactId>
             <version>1.1.7</version>
