+++ metrics-reporter-influxdb/src/main/java/org/graylog/plugins/metrics/influxdb/MetricsInfluxDbReporterConfiguration.java
@@ -21,7 +21,7 @@ import com.github.joschi.jadconfig.ValidationException;
 import com.github.joschi.jadconfig.ValidatorMethod;
 import com.github.joschi.jadconfig.util.Duration;
 import com.github.joschi.jadconfig.validators.PositiveDurationValidator;
-import org.graylog.plugins.metrics.influxdb.jadconfig.StringMapConverter;
+import org.graylog.plugins.metrics.core.jadconfig.StringMapConverter;
 import org.graylog2.plugin.PluginConfigBean;
 
 import java.net.URI;
