+++ metrics-reporter-mongodb/pom.xml
@@ -19,6 +19,11 @@
 
     <dependencies>
         <dependency>
+            <groupId>org.graylog.plugins</groupId>
+            <artifactId>metrics-reporter-core</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+        <dependency>
             <groupId>io.github.aparnachaudhary</groupId>
             <artifactId>mongodb-metrics</artifactId>
             <version>0.0.2</version>
